requirements.txt
requests
python-dateutil
python-dotenv
flask
gunicorn
